requirements.txt
streamlit==1.32.2
streamlit-modal==0.1.2
streamlit-option-menu==0.3.12
streamlit-server-state==0.17.1
pandas==2.1.4
requests==2.31.0
Pillow==10.1.0
duckduckgo_search==5.3.0